Tweak installation instructions
Originally submitted with a wget to a user's raw directory. Security weakness here is that arbitrary code could have been inserted by user. The contributor seemed nice enough, but I'd be concerned about having that in the repo. Too easy for someone to do a wget without examining the file, assuming it would be the same as contained on this repo.